Adding Inspire validated version of resolved example
This makes a number of changes to comply with tests ran by the Inspire online validator.

Some changes have been made based on technical recommendations in the relevent Technical Guidance documents, on which the validator tests are based, but some aspects are not directly tested, but we have incorporated.

Once complete these changes should be merged into the resolved example.