Adding sample record for UK PDC Discovery profile
This replaces the previous candidate record with one that complies with the UK PDC Discovery profile.

This record is largely the same but is identified as being a test record, and no longer refers to any existing record, but rather pretends to be something typical. It has a test DOI to allow for identifiers and required citations to be used, and links to the test dataset artefact as a transfer option.