Replacing Python Flask app with Jekyll static site
Dynamic record generation removed in favour of using an existing record as an example - we couldn’t do this before as there weren’t any example records to use.

Removed RDF vocab processing for displaying GCMD terms as there is now a proper viewer and the cost of doing this from a build time/complexity perspective was not worth it.

Candidate records for other standards removed until they can be implemented properly.